add assets to doc images
