--- lib/widget/theme_Provider.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ThemeProvider extends ChangeNotifier{
-  ThemeMode themeMode= ThemeMode.light;
-  bool get isDarkMode=>themeMode==ThemeMode.dark;
-
-  void toggleTheme(bool ison){
-    themeMode=ison?ThemeMode.dark:ThemeMode.light;
-    notifyListeners();
-  }
-}
-class MyThemes{
-  static final darkTheme=ThemeData(
-    scaffoldBackgroundColor: Colors.grey.shade900,
-    colorScheme: ColorScheme.dark(),
-  );
-  
-  static final lightTheme=ThemeData(
-    scaffoldBackgroundColor: Colors.white,
-    colorScheme: ColorScheme.light(),
-  );
-}
